refactor(tree): refactor tree component
