+++ package.json
@@ -41,6 +41,7 @@
     "@mantine/modals": "6.0.7",
     "@mantine/notifications": "5.9.5",
     "@mantine/spotlight": "5.9.5",
+    "@minoru/react-dnd-treeview": "^3.4.4",
     "@monaco-editor/react": "4.4.6",
     "@next/bundle-analyzer": "12.3.1",
     "@next/font": "^13.4.3",
@@ -113,6 +114,7 @@
     "react-contextmenu": "2.14.0",
     "react-csv-reader": "4.0.0",
     "react-dev-inspector": "1.8.4",
+    "react-dnd": "^16.0.1",
     "react-dom": "18.2.0",
     "react-draggable": "4.4.5",
     "react-dropzone": "14.2.3",
