+++ src/components/Tree/fileIcon.tsx
@@ -1,0 +1,48 @@
+import { useStore } from '@/store/index';
+import { FilesItemType, VSCodeRemoteFolderName } from '@/store/lib/w3bstream/schema/filesList';
+import { Image, ImageProps, Box, Flex, Portal, Text, Tooltip, Divider, Center, Spinner } from '@chakra-ui/react';
+
+export const FileIcon = (file: FilesItemType) => {
+  const {
+    w3s: { projectManager }
+  } = useStore();
+  //https://github.com/PKief/vscode-material-icon-theme/tree/main/icons
+  const s: ImageProps = {
+    h: 5,
+    w: 5,
+    mr: 1
+  };
+  if (file?.label?.endsWith('.go')) {
+    return <Image {...s} src="/images/icons/go.svg"></Image>;
+  } else if (file?.label?.endsWith('.html')) {
+    return <Image {...s} src="/images/icons/html.svg"></Image>;
+  } else if (file?.label?.endsWith('.ts')) {
+    return <Image {...s} src="/images/icons/typescript.svg"></Image>;
+  } else if (file?.label?.endsWith('.wasm')) {
+    return <Image {...s} src="/images/icons/assembly.svg"></Image>;
+  } else if (file?.label?.endsWith('.flow')) {
+    return <Image {...s} src="/images/icons/tree.svg"></Image>;
+  } else if (file?.label?.endsWith('.json')) {
+    return <Image {...s} src="/images/icons/json.svg"></Image>;
+  } else if (file?.label == '.env') {
+    return <Image {...s} src="/images/icons/settings.svg"></Image>;
+  }
+
+  if (file?.type == 'file') {
+    return <Image {...s} src="/images/icons/file.svg"></Image>;
+  }
+
+  if (file?.label == 'Browser Files') {
+    return <>{file.isOpen ? <Image {...s} src="/images/icons/folder-client-open.svg"></Image> : <Image {...s} src="/images/icons/folder-client.svg"></Image>}</>;
+  } else if (file?.label == VSCodeRemoteFolderName) {
+    return (
+      <>
+        {file.isOpen ? (
+          <Image {...s} src="/images/icons/folder-vscode-open.svg" filter={!projectManager.isWSConnect ? 'grayscale(100%)' : ''}></Image>
+        ) : (
+          <Image filter={!projectManager.isWSConnect ? 'grayscale(100%)' : ''} {...s} src="/images/icons/folder-vscode.svg"></Image>
+        )}
+      </>
+    );
+  }
+};
