+++ src/pages/_app.tsx
@@ -19,7 +19,6 @@ import { WagmiProvider } from '@/components/WagmiProvider';
 import '@/lib/superjson';
 import { Global, css } from '@emotion/react';
 import { Inter } from '@next/font/google';
-
 const oxanium = Inter({ subsets: ['latin'] });
