+++ src/pages/_document.tsx
@@ -8,8 +8,15 @@ export default class _Document extends Document {
           {/* <link href="https://fonts.googleapis.com/css2?family=Oxanium&display=swap" rel="stylesheet" /> */}
           <link rel="stylesheet" href="https://unpkg.com/primeicons/primeicons.css" />
           <link rel="icon" href="favicon.svg" sizes="any" type="image/svg+xml" />
+          <style>{`
+            ul {
+              list-style: none;
+              margin: 0;
+              padding: 0;
+            }
+          `}</style>
         </Head>
-        <body >
+        <body>
           <Main />
           <NextScript />
         </body>
